chore(oxlint): enable no-unnecessary-type-assertion, clear the backlog
Restore typescript/no-unnecessary-type-assertion to error (parked at off in the
tsgolint 7 upgrade, #17720). oxlint --fix removes ~1350 genuinely-unnecessary
assertions across ~400 files.

The remaining ~130 flags are tsgolint-7-vs-typescript-7 false positives — the
new 'receiver accepts original type' sub-check flags assertions that tsc@7 still
requires (partial test fixtures asserted to full types, 'as never' escape
hatches). Removing them breaks tsc, so the 88 affected files carry a file-level
// oxlint-disable with a rationale pointing at #17746 rather than papering over
them per line. Closes #17746.

Verified: tsc typecheck:node + typecheck:web pass (0 errors); oxlint reports 0
no-unnecessary-type-assertion violations and 0 errors; oxfmt clean. The full
test suite runs on CI (better-sqlite3 ABI + memory prevent a local run).

+++ src/main/services/file/utils/content.ts
@@ -67,7 +67,7 @@ export async function readByPath(
 
     if (isSameVersion(before, after) && (readByteLength === undefined || after.size === readByteLength)) {
       if (encoding === 'binary' && (options as BinaryReadOptions).withContentHash) {
-        return { content, mime: contentMime, version: after, contentHash: hashContent(content as Uint8Array) }
+        return { content, mime: contentMime, version: after, contentHash: hashContent(content) }
       }
       return { content, mime: contentMime, version: after }
     }
